feat: git storage transaction semaphore implementation
** What is the motivation for the change?

We want to be able to lock the access to the storage to ensure updates
are performed atomically.

For the Git storage, this will be done through semaphores, an
implementation was added in this commit.

** Others changes

- We no longer pass the storage directly to handler, instead we give
- Clippy and rustfmt

** Breaking changes

The InMemory storage doesn't have a transaction implementation, using it
will result in a 'todo!' error