feat(core): keep an history of all tasks in git storage
** What is the motivation for the change?

We want to keep an history of every task to allow commands such as logs
and restart in the CLI.

** How does it address the problem?

We keep a list of all logs in the `__metadata.json` file inside the git
storage.

** Are there any side effects?

This will clutter the metadata file with a lot of infos, but that seems
fine, for example, 20 tasks a day for a whole year will use about 0.2 MB
of space.

The real issue could come from either:
- the conversion of the ulid to timestamps
- opening a large amount of json files to get the logs infos