feat(CLI): improve log messages formatting
** What is the motivation for the change?

I want a better way to visualize tasks in the CLI.

** How does it address the problem?

We now have three output options:

A short version that we get by default that omit a lot of informations
tasks (omit ID, task end date ...).

A detailed version when running with `--verbose` that display
informations in a table format.

A json version when running with `--json` that display the tasks as a
json array.