feat: git synchronization, storage refactor, config profiles, CLI
** What is the motivation for the change?

This is a mega-commit that introduce many changes to the CLI

Git storage synchronization:

We can now synchronize the data accross multiple devices, we use a
checkpoint-based approach for reconciliation.

This is the first basic working version of the git storage but many
changes and optimizations will be added in future commits.

Configuration profile:

We can now have multiple profiles in the configuration file, an option
was added to the CLI to select the profile we wish to use.

Better CLI log:

Some changes were added to the CLI to better display the tasks logs, we
can now choose between a simplified, table or json display.

Packages refactors:

We now have three storage packages, they are used as followed

- storage-core: contains the core structure that different storage type
needs to implement.

- storage: interface that should be used by the different bricks of the
applications (*ongoing update*)

- storage-git: first storage implementation

... We also now have a config package which alleviate a lot of issue I had
with cross dependencies.