build: automatic package versionning
** Reason for the change

This commit automate the creation of the versio configuration file which
means we no longer need to update the `versio-projects.json` file at the
root of the project (this was always a pain point when adding packages).

** Side-effects

Deletion of all `CHANGELOG.html` files, since all projects ids changed
we needed to ensure that their changelog files were resetted to avoid
versio hallucination.