feat(core): hash value instead of a versionning number on task change
We previously used a `__version` field on tasks to verify whether they
have been updated. We used this field in the frontend to know the
version of an object to know when a re-render need to occur.

Although this works fine, I can see how the `__version` field can be
misleading, mainly, we could mistake it for the document versioning
number. Using a `__hash` field instead, which contains the hash of the
current object seems like a more elegant solution.