feat: trie index for task ids
Implemented a trie index to compute tasks prefix ids; to get the minimum
unique identifier for a task, consider the following task ids and their
prefixes:

abc123 → abc12
abc456 → abc4
abc → abc
aaa → aa
abc111 → abc11

If we were to add “abc444” we need to properly modify previous id to
match, here:
abc456 → abc45
abc444 → abc44

This is the purpose of the trie structure. The trie structure is fast on
read slow on write, it’s composed of a mix of disk storage (native-db) &
memory storage that caches nodes on read (dashmap).

NB: there’s no eviction policy for the in-memory cache, we may revisit
this but the benchmark using “moka” and “lfu” induced a huge overhead:
• Rwlock<hashmap> ~ 90ms
• dashmap ~ 70ms
• moka or lfu ~ 450ms
• native-db only ~ 1.7 sec

Also on this commit:
• feat: added a command to query native-db directly from the CLI for debugging
• feat: can now provide an id to the resume command (previously restart)
• feat: improved CLI error message on D-Bus connection error
• fix: added a ping command so that CLI proxy connection to D-Bus
endpoint is verified accordingly before jumping to a command.