feat: introduce project coloring
This commit introduce a way to differentiate project by color. This will
be used by cli for coloring (but is not used in this commit).

Creating a task with a unique project name will auto-assign the project
a color, following tasks sharing this project name won’t have to repeat
this action.

• project colors are randomized but I added some incentive to use colors
that are further away.
• no way to modify a project color in this commit