feat: better input for date and time format in CLI commands
It was sort of hard to set the date easily while using the CLI, for
example what if I need to set the end date to yesterday at 3pm? Or on a
precise day at a precise time?

Also; What if I wanted to set a task duration time instead of a date to
avoid mental calculation?