feat: rename start command and accept custom date
Small scenario before this commit; If we wanted to create a task at a
specific time, for example if we forgot to log some task, we needed to
start a new task (which would stop the current) and then edit it.

Now we can create a new task while leaving the current task running, we
renamed the 'start' command to 'new' for clarity.

We also show conflict in logs output.